pyproject.toml
[tool.poetry]
name = "course_work_drf"
version = "0.1.0"
description = ""
authors = ["KirillOreshchuk <oreshchuk.kirill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
coverage = "^7.4.1"
python-dotenv = "^1.0.1"
celery = "^5.3.6"
redis = "^5.0.1"
requests = "^2.31.0"
flake8 = "^7.0.0"
django-celery-beat = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
